requirements.txt
fastapi-jwt-auth2>=0.5.0,<1.0.0
fastapi>=0.96,<1.0
pydantic>=2.0.0,<2.2.0
uvicorn>=0.22.0, <1.0.0
pyjwt>=1.7.1,<2.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0
python-jose>=3.3.0,<4.0.0
databases>=0.7.0<1.0.0
sqlalchemy
psycopg2-binary>=2.9.1,<3.0.0
asyncpg>=0.28.0,<1.0.0
httpx>=0.19.0,<1.0.0
bcrypt>=4.0.1,<5.0.0
tortoise-orm>=0.19.3,<1.0.0
PyYAML>=6.0c
slowapi>=0.1.8, <1.0.0
google-api-python-client>=2.87.0,<3.0.0
fastapi-jwt-auth2>=0.5.0,<1.0.0
python-social-auth>=0.3.6,<1.0.0
Authlib>=1.2.1,<2.0.0
itsdangerous>=2.1.2,<3.0.0
httpx >=0.24.1, <1.0.0
fastapi-sso>=0.7.2, <1.0.0
Jinja2>=3.1.2,<4.0.0